Remove person flag on module type
Given there is no real harm in sending extra attributes
to the API that aren't processed, just have the names
all sent together. If it turns out to be a Person type
then it will use the first_name, last_name fields.
Otherwise it will use the sent name field.